Corrected the IP Address in the DNS Zone and the hosts File. They are pointing to the wrong server.
Add the Certificate from the LabCA to the
TrustedRootCertificate Authorities.
We have to do that because the certificate of the RDS deployment is not trusted by the local computer.

Added Manifest File for the SetupHelper Module.
Modified the Connection Authorization Policy on the GatewayServer
and add the right group to It.

Removed the Resource Authorization Policies and created my own one.
Add the appropriate Group to It.